fix: add FromPrimitiveOrF16 to NativePType replacing `FromPrimitive`
