+++ vortex-dtype/src/lib.rs
@@ -18,6 +18,7 @@ pub mod datetime;
 mod decimal;
 mod dtype;
 mod extension;
+mod f16;
 mod field;
 mod field_mask;
 mod field_names;
@@ -31,6 +32,7 @@ pub use bigint::*;
 pub use decimal::*;
 pub use dtype::{DType, NativeDType};
 pub use extension::*;
+pub use f16::*;
 pub use field::*;
 pub use field_mask::*;
 pub use field_names::*;
